Added column definition to the base package

+++ helpers/aggs.go
@@ -18,22 +18,22 @@ func MaxIntSlice(nums []int) int {
 
 // MaxSeries returns the maximum of a series
 func MaxSeries(num []interface{}, dtype base.DType) interface{} {
-	maxInt := int64(math.MinInt64)
-	maxFloat := float64(math.MinInt64)
+	maxInt := math.MinInt64
+	maxFloat := float32(math.MinInt64)
 
 	switch dtype {
 	case base.Int:
 		for _, val := range num {
-			if val.(int64) > maxInt {
-				maxInt = val.(int64)
+			if val.(int) > maxInt {
+				maxInt = val.(int)
 			}
 		}
 		return maxInt
 
 	case base.Float:
 		for _, val := range num {
-			if val.(float64) > maxFloat {
-				maxFloat = val.(float64)
+			if val.(float32) > maxFloat {
+				maxFloat = val.(float32)
 			}
 		}
 		return maxFloat
@@ -43,22 +43,22 @@ func MaxSeries(num []interface{}, dtype base.DType) interface{} {
 
 // MinSeries returns the minimum of a series
 func MinSeries(num []interface{}, dtype base.DType) interface{} {
-	minInt := int64(math.MaxInt64)
-	minFloat := float64(math.MaxInt64)
+	minInt := math.MaxInt64
+	minFloat := float32(math.MaxInt64)
 
 	switch dtype {
 	case base.Int:
 		for _, val := range num {
-			if val.(int64) < minInt {
-				minInt = val.(int64)
+			if val.(int) < minInt {
+				minInt = val.(int)
 			}
 		}
 		return minInt
 
 	case base.Float:
 		for _, val := range num {
-			if val.(float64) < minFloat {
-				minFloat = val.(float64)
+			if val.(float32) < minFloat {
+				minFloat = val.(float32)
 			}
 		}
 		return minFloat
